requirements_priority3.txt
# Priority 3 Machine Learning Dependencies
# Add these to requirements.txt for Priority 3 implementation

# Core ML Libraries
scikit-learn==1.4.0
xgboost==2.0.3
numpy==1.26.3
pandas==2.2.0

# Feature Engineering and Text Processing
nltk==3.8.1
textstat==0.7.3
sentence-transformers==2.2.2

# Caching and Performance
cachetools==5.3.2
joblib==1.3.2
multiprocessing-logging==0.3.4

# Memory Optimization
psutil==5.9.8  # Already in requirements.txt
memory-profiler==0.61.0

# Model Persistence and Versioning
# pickle5 removed - not needed in Python 3.8+, using built-in pickle
dill==0.3.7

# Advanced Analytics
plotly==5.17.0
seaborn==0.13.1
matplotlib==3.8.2

# Configuration and Utilities
pyyaml==6.0.1  # Already in requirements.txt
python-dateutil==2.8.2
tqdm==4.66.1

# Optional: Deep Learning (if needed for advanced features)
# torch==2.1.2
# transformers==4.36.2

# Development and Testing (Priority 3 specific)
pytest-benchmark==4.0.0
pytest-mock==3.12.0
hypothesis==6.92.1

# Monitoring and Metrics
prometheus-client==0.19.0
grafana-api==1.0.3

# Multi-language Support - removed problematic packages
# polyglot==16.07.04  # Has complex dependencies
# pycld2==0.41       # Has compilation issues on Windows

# Performance Profiling
line-profiler==4.1.1
# py-spy==0.3.14  # May have platform issues 